requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-docx
pandas
